Allow to not delete the service folder using property
Using `-Dts.global.delete.folder.on.exit=false`

+++ quarkus-test-core/src/main/java/io/quarkus/test/bootstrap/BaseService.java
@@ -28,6 +28,7 @@ public class BaseService<T extends Service> implements Service {
 
     public static final String SERVICE_STARTUP_TIMEOUT = "startup.timeout";
     public static final Duration SERVICE_STARTUP_TIMEOUT_DEFAULT = Duration.ofMinutes(5);
+    public static final String DELETE_FOLDER_ON_EXIT = "delete.folder.on.exit";
 
     private static final String SERVICE_STARTUP_CHECK_POLL_INTERVAL = "startup.check-poll-interval";
     private static final Duration SERVICE_STARTUP_CHECK_POLL_INTERVAL_DEFAULT = Duration.ofSeconds(2);
@@ -210,10 +211,12 @@ public class BaseService<T extends Service> implements Service {
     @Override
     public void close() {
         stop();
-        try {
-            FileUtils.deletePath(getServiceFolder());
-        } catch (Exception ex) {
-            Log.warn(this, "Could not delete service folder. Caused by " + ex.getMessage());
+        if (getConfiguration().isTrue(DELETE_FOLDER_ON_EXIT)) {
+            try {
+                FileUtils.deletePath(getServiceFolder());
+            } catch (Exception ex) {
+                Log.warn(this, "Could not delete service folder. Caused by " + ex.getMessage());
+            }
         }
     }
